+++ log/topic.go
@@ -709,6 +709,40 @@ func internalTopic(l Logger, d trace.Detailer) (t trace.Topic) {
 			}
 		}
 	}
+	t.OnWriterBeforeCommitTransaction = func(
+		info trace.TopicOnWriterBeforeCommitTransactionStartInfo,
+	) func(trace.TopicOnWriterBeforeCommitTransactionDoneInfo) {
+		if d.Details()&trace.TopicWriterStreamLifeCycleEvents == 0 {
+			return nil
+		}
+
+		start := time.Now()
+
+		return func(doneInfo trace.TopicOnWriterBeforeCommitTransactionDoneInfo) {
+			ctx := with(*info.Ctx, TRACE, "ydb", "topic", "writer", "beforecommit")
+			l.Log(ctx, "wait of flush messages before commit transaction",
+				String("kqp_session_id", info.KqpSessionID),
+				String("topic_session_id_start", info.TopicSessionID),
+				String("topic_session_id_finish", doneInfo.TopicSessionID),
+				String("tx_id", info.TransactionID),
+				latencyField(start),
+			)
+		}
+	}
+	t.OnWriterAfterFinishTransaction = func(
+		info trace.TopicOnWriterAfterFinishTransactionStartInfo,
+	) func(trace.TopicOnWriterAfterFinishTransactionDoneInfo) {
+		start := time.Now()
+
+		return func(doneInfo trace.TopicOnWriterAfterFinishTransactionDoneInfo) {
+			ctx := with(context.Background(), TRACE, "ydb", "topic", "writer", "beforecommit")
+			l.Log(ctx, "close writer after transaction finished",
+				String("kqp_session_id", info.SessionID),
+				String("tx_id", info.TransactionID),
+				latencyField(start),
+			)
+		}
+	}
 	t.OnWriterClose = func(info trace.TopicWriterCloseStartInfo) func(doneInfo trace.TopicWriterCloseDoneInfo) {
 		if d.Details()&trace.TopicWriterStreamLifeCycleEvents == 0 {
 			return nil
@@ -835,6 +869,7 @@ func internalTopic(l Logger, d trace.Detailer) (t trace.Topic) {
 			Int64("written_offset_min", acks.WrittenOffsetMin),
 			Int64("written_offset_max", acks.WrittenOffsetMax),
 			Int("written_offset_count", acks.WrittenCount),
+			Int("written_in_tx_count", acks.WrittenInTxCount),
 			Int("skip_count", acks.SkipCount),
 			versionField(),
 		)