+++ topic/topicwriter/topicwriter.go
@@ -10,12 +10,12 @@ type (
 	Message = topicwriterinternal.PublicMessage
 )
 
-var ErrQueueLimitExceed = topicwriterinternal.PublicErrQueueIsFull
+var ErrMessagesPutToInternalQueueBeforeError = topicwriterinternal.PublicErrMessagesPutToInternalQueueBeforeError
 
 // Writer represent write session to topic
 // It handles connection problems, reconnect to server when need and resend buffered messages
 type Writer struct {
-	inner *topicwriterinternal.Writer
+	inner *topicwriterinternal.WriterReconnector
 }
 
 // PublicInitialInfo is an information about writer after initialize
@@ -24,7 +24,7 @@ type PublicInitialInfo struct {
 }
 
 // NewWriter create new writer from internal type. Used internally only.
-func NewWriter(writer *topicwriterinternal.Writer) *Writer {
+func NewWriter(writer *topicwriterinternal.WriterReconnector) *Writer {
 	return &Writer{
 		inner: writer,
 	}
@@ -38,9 +38,11 @@ func NewWriter(writer *topicwriterinternal.Writer) *Writer {
 // especially when connection has problems.
 //
 // It returns ErrQueueLimitExceed (must be checked by errors.Is)
-// if ctx cancelled before messages put to internal buffer or try to add more messages, that can be put to queue
+// if ctx cancelled before messages put to internal buffer or try to add more messages, that can be put to queue.
+// If err != nil you can check errors.Is(err, ErrMessagesPutToInternalQueueBeforeError) for check if the messages
+// put to buffer before error. It means that it is messages can be delivered to the server.
 func (w *Writer) Write(ctx context.Context, messages ...Message) error {
-	return w.inner.Write(ctx, messages...)
+	return w.inner.Write(ctx, messages)
 }
 
 // WaitInit waits until the reader is initialized
@@ -75,4 +77,52 @@ func (w *Writer) Close(ctx context.Context) error {
 // Flush waits till all in-flight messages are acknowledged.
 func (w *Writer) Flush(ctx context.Context) error {
 	return w.inner.Flush(ctx)
+}
+
+// TxWriter used for send messages to the transaction
+//
+// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
+type TxWriter struct {
+	inner *topicwriterinternal.WriterWithTransaction
+}
+
+func NewTxWriterInternal(w *topicwriterinternal.WriterWithTransaction) *TxWriter {
+	return &TxWriter{inner: w}
+}
+
+// Write messages to the transaction
+//
+// It has not retries. If fails - needs to retry full transaction, as with any other
+// error with table.
+//
+// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
+func (w *TxWriter) Write(ctx context.Context, messages ...Message) error {
+	return w.inner.Write(ctx, messages...)
+}
+
+// WaitInit waits until the reader is initialized
+// or an error occurs, return PublicInitialInfo and err
+//
+// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
+func (w *TxWriter) WaitInit(ctx context.Context) (err error) {
+	_, err = w.inner.WaitInit(ctx)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// WaitInitInfo waits until the reader is initialized
+// or an error occurs, return PublicInitialInfo and err
+//
+// Experimental: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#experimental
+func (w *TxWriter) WaitInitInfo(ctx context.Context) (info PublicInitialInfo, err error) {
+	privateInfo, err := w.inner.WaitInit(ctx)
+	if err != nil {
+		return PublicInitialInfo{}, err
+	}
+	publicInfo := PublicInitialInfo{LastSeqNum: privateInfo.LastSeqNum}
+
+	return publicInfo, nil
 }